third_party/blink/web_tests/platform/generic/fast/events/tab-test-not-visible-imagemap-expected.txt





Test that tabbing does not focus area element when its associated image is not rendered.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement.id is 'input1'
PASS document.activeElement.id is 'input2'
PASS document.activeElement.id is 'input3'
PASS successfullyParsed is true

TEST COMPLETE
